Update Controller.java
I am not sure about the print method, we are suppose to print out the outstate and the instate according to the button we choose, but first should we do a search to pick out the things we should print? Second, the type of the print method is Student but the outputArea is suppose to print out strings,should I change the print method in the student list or do something else.